Change enum of Scala 2 to be 3 for Topping enum

+++ PizzaOopExample/src/main/scala/pizza/Pizza.scala
@@ -22,7 +22,7 @@ case class Pizza(
   }
 
   def getPrice(
-    toppingsPrices: Map[Topping, Int],
-    crustSizePrices: Map[CrustSize, Int],
-    crustTypePrices: Map[CrustType, Int]
+    toppingsPrices: Map[Topping, Double],
+    crustSizePrices: Map[CrustSize, Double],
+    crustTypePrices: Map[CrustType, Double]
   ): Int = ???
